Set :site client option to 'https://auth.exvo.com'
In other words, make it a fixed string. Also add a note in the README
about overriding it in the omniauth initializer.

The problem with the previous implementation using the exvo_helpers
gem was that it was impossible to override the auth host setting using
the exvo_helpers gem, because omniauth-exvo gem would not pick it up.

That's why it's a fixed string now and the dependency on the
exvo_helpers gem has been dropped.

Also tag a new 0.0.3 version.